eso_tracker 1st version before optimizing realtime performance

+++ src/mapping/lio-sam/include/ivox3d/ivox3d.h
@@ -354,7 +354,14 @@ std::vector<float> IVox<dim, node_type, PointType>::StatGridPoints() const
   }
   float ave    = float( sum ) / num;
   float stddev = num > 1 ? sqrt( ( float( sum_square ) - num * ave * ave ) / ( num - 1 ) ) : 0;
-  return std::vector<float>{ valid_num, ave, max, min, stddev };
+  //return std::vector<float>{ valid_num, ave, max, min, stddev };
+  return std::vector<float>{ 
+    static_cast<float>(valid_num), 
+    ave, 
+    static_cast<float>(max), 
+    static_cast<float>(min), 
+    stddev 
+};
 }
 
 }  // namespace faster_lio
